Created a script that generates openAPI docs for httpTrigger.ts
