--- packages/backend/experimental/docsGenerator/test.sh
@@ -1,4 +1,0 @@
-#!/usr/bin/env bash
-
-npx tsx translateAzureToSwagger.ts httpTrigger.ts foo.yaml
-
