+++ packages/backend/scripts/experimental/docsGenerator/install-dependencies.sh
