+++ packages/backend/scripts/experimental/docsGenerator/test.sh
@@ -1,0 +1,5 @@
+#!/usr/bin/env bash
+
+npx tsx experimental/docsGenerator/translateAzureToSwagger.ts experimental/docsGenerator/httpTrigger.ts \
+    ../api-docs/openAPI-docs.yaml
+
