+++ packages/backend/scripts/generate-docs.sh
@@ -1,0 +1,4 @@
+#!/usr/bin/env bash
+
+npx tsx experimental/docsGenerator/translateAzureToSwagger.ts ../src/functions/httpTrigger.ts \
+    ../api-docs/openAPI-docs.yaml
